+++ src/EvDevSharp/Enums/EvDevGuessedDeviceType.cs
@@ -1,0 +1,12 @@
+namespace EvDevSharp;
+
+public enum EvDevGuessedDeviceType
+{
+    Unknown,
+    Keyboard,
+    Mouse,
+    TouchPad,
+    TouchScreen,
+    Tablet,
+    GamePad
+}
