+++ src/EvDevSharp/Enums/EvDevKeyValue.cs
@@ -1,0 +1,8 @@
+namespace EvDevSharp;
+
+public enum EvDevKeyValue
+{
+    KeyUp = 0,
+    KeyDown = 1,
+    KeyRepeat = 2
+}
