+++ src/EvDevSharp/EventArgs/EvDevEventArgs.cs
@@ -1,0 +1,10 @@
+namespace EvDevSharp;
+
+public class EvDevEventArgs : EventArgs
+{
+    public EvDevEventArgs(int code, int value) =>
+        (Code, Value) = (code, value);
+
+    public int Code { get; set; }
+    public int Value { get; set; }
+}
