+++ src/EvDevSharp/EventArgs/OnRelativeEventArgs.cs
@@ -1,0 +1,10 @@
+namespace EvDevSharp;
+
+public class OnRelativeEventArgs : EventArgs
+{
+    public OnRelativeEventArgs(EvDevRelativeAxisCode axis, int value) =>
+        (Axis, Value) = (axis, value);
+
+    public EvDevRelativeAxisCode Axis { get; set; }
+    public int Value { get; set; }
+}
