+++ src/EvDevSharp/EventArgs/OnSynEventArgs.cs
@@ -1,0 +1,10 @@
+namespace EvDevSharp;
+
+public class OnSynEventArgs : EventArgs
+{
+    public OnSynEventArgs(EvDevSynCode code, int value) =>
+        (Code, Value) = (code, value);
+
+    public EvDevSynCode Code { get; set; }
+    public int Value { get; set; }
+}
