+++ src/EvDevSharp/InteropStructs/EvDevAbsAxisInfo.cs
@@ -1,0 +1,14 @@
+using System.Runtime.InteropServices;
+
+namespace EvDevSharp.InteropStructs;
+
+[StructLayout(LayoutKind.Sequential, Pack = 4)]
+public struct EvDevAbsAxisInfo
+{
+    public int Value;
+    public int Min;
+    public int Max;
+    public int Fuzz;
+    public int Flat;
+    public int Resolution;
+}
