+++ src/EvDevSharp/InteropStructs/TimeVal.cs
@@ -1,0 +1,10 @@
+using System.Runtime.InteropServices;
+
+namespace EvDevSharp.InteropStructs;
+
+[StructLayout(LayoutKind.Sequential)]
+internal struct TimeVal
+{
+    public nint tv_sec;
+    public nint tv_usec;
+}
